[#25] test: Create a working test
